test: address TOON review comments
